separate out MenuItem and Icon styled components
