+++ src/components/Toolbar/SaveFile/index.tsx
@@ -3,8 +3,9 @@ import { saveAs } from "file-saver";
 
 import { useMarkdownState } from "contexts/MarkdownContext";
 
-import SaveIcon from "icons/save.svg";
-import { MenuItem, MenuItemWithTooltip } from "../MenuItem";
+import SaveSvg from "icons/save.svg";
+import { MenuItem } from "../MenuItem";
+import { IconWithTooltip } from "../Icon";
 import { Tooltip } from "../Tooltip";
 import { SaveFileForm } from "./SaveFileForm";
 
@@ -34,14 +35,12 @@ const SaveFileMenuItem: React.FC = () => {
       />
     </MenuItem>
   ) : (
-    <MenuItemWithTooltip
-      aria-label="save-file"
-      onClick={openForm}
-      role="button"
-    >
-      <SaveIcon />
-      <Tooltip text="SAVE" />
-    </MenuItemWithTooltip>
+    <MenuItem aria-label="save-file" onClick={openForm} role="button">
+      <IconWithTooltip>
+        <SaveSvg />
+        <Tooltip text="SAVE" />
+      </IconWithTooltip>
+    </MenuItem>
   );
 };
